Reorganize file structure and readme

+++ ricks_config_tool.sh
@@ -4,26 +4,6 @@
 # and configure LAMP stack to run a simple PHP web app.
 # See included README for instructions on usage and configuration.
 
-
-# Populate our install_list array
-declare -a install_list
-while IFS='\n' read -r value; do
-  $value >> install_list
-done < "packages/install.txt"
-
-# Populate our removal_list array
-declare -a removal_list
-while IFS='\n' read -r value; do
-  $value >> removal_list
-done < "packages/uninstall.txt"
-
-# Build our associative array from key val pairs in metadata.txt
-# Note: Syntax and order are important in this text file
-declare -A metadata
-while IFS== read -r key value; do
-  metadata[$key]=$value
-done < "metadata.txt"
-
 # Check for package manager updates and upgrade to latest
 sudo apt-get update && sudo apt-get upgrade
 
@@ -35,15 +15,26 @@ sudo apt-get update && sudo apt-get upgrade
 #   "libapache2-mod-php5"
 #   "php5-mcrypt"
 # )
-# Iterate removal_list. Remove if not already removed.
-for rm_pkg in "${removal_list[@]}"
-do
-  if dpkg -l | grep -i "${rm_pkg}"; then
-    sudo apt-get purge -y "${rm_pkg}" # Remove package and associated config files
-    sudo apt-get autoremove -y # Remove packages installed by other packages that aren't needed
-    sudo apt-get autoclean -y # Removes .deb files no longer installed.
-  fi
-done
+
+# Runs only if uninstall list is not empty
+if [ -s debian_packages/uninstall.txt ]; then
+  # Populate our removal_list array
+  declare -a removal_list
+  while IFS='\n' read -r value; do
+    removal_list+=( "${value}" )
+  done < "debian_packages/uninstall.txt"
+
+  # Iterate removal_list. Remove if not already removed.
+  for rm_pkg in "${removal_list[@]}"
+  do
+    if dpkg -l | grep -i "${rm_pkg}"; then
+      sudo apt-get remove "${rm_pkg}" # Remove package
+      sudo apt-get autoremove # Remove packages installed by other packages that aren't needed
+      sudo apt-get purge -y $(dpkg --list |grep '^rc' |awk '{print $2}') # Remove package config files
+      sudo apt-get clean # Removes .deb files no longer installed.
+    fi
+  done
+fi
 
 # --------------------------
 # Install Debian Packages
@@ -57,20 +48,33 @@ done
 #   "php5-mcrypt"
 # )
 
-# Iterate install_list array. Install if not already installed.
-for in_pkg in "${install_list[@]}"
-do
-  if ! dpkg -l | grep "${in_pkg}"; then
-    sudo apt-get install -y "${in_pkg}"
-    if [ "${in_pkg}" == "mysql-server" ]; then
-      # Create the MySQL database
-      sudo mysql_install_db
+# Runs only if install list is not empty
+if [ -s debian_packages/install.txt ]; then
+  # Populate our install_list array
+  declare -a install_list
+  while IFS='\n' read -r value; do
+    install_list+=( "${value}" )
+  done < "debian_packages/install.txt"
 
-      # Secure our installation by removing insecure defaults
-      sudo mysql_secure_installation
+  # Iterate install_list array. Install if not already installed.
+  for in_pkg in "${install_list[@]}"
+  do
+    if ! dpkg -l | grep "${in_pkg}"; then
+      sudo apt-get install -y "${in_pkg}"
+      if [ "${in_pkg}" == "mysql-server" ]; then
+        # Create the MySQL database
+        sudo mysql_install_db
+
+        # Secure our installation by removing insecure defaults
+        sudo mysql_secure_installation
+      fi
     fi
-  fi
-done
+  done
+fi
+
+# --------------------------
+# Setting Metadata and Content
+# --------------------------
 
 # Replace default index.html with index.php if the file exists
 if [ -f "/var/www/html/index.html" ]; then
@@ -78,9 +82,20 @@ if [ -f "/var/www/html/index.html" ]; then
   touch /var/www/html/index.php
 fi
 
-# --------------------------
-# Setting Metadata and Content
-# --------------------------
+# Build our associative array from key val pairs in metadata.txt
+# Note: Syntax and order are important in this text file
+
+# Runs only if uninstall list is not empty
+if [ -s metadata.txt ]; then
+  declare -A metadata
+  while IFS== read -r key value; do
+    metadata[$key]=$value
+  done < "metadata.txt"
+  sudo chmod "${metadata[permissions]}" "${metadata[file]}"
+  sudo chown "${metadata[owner]}" "${metadata[file]}"
+  sudo chgrp "${metadata[group]}" "${metadata[file]}"
+  sudo echo "${metadata[content]}" > "${metadata[file]}"
+fi
 
 # To change metadata for files or folders, use the following example as a guide.
 
@@ -92,10 +107,10 @@ fi
 #   ["file"]="/var/www/html/index.php"
 #   ["content"]='<?php header("Content-Type: text/plain"); echo "Hello, world!\n";'
 # )
-sudo chmod "${metadata[permissions]}" "${metadata[file]}"
-sudo chown "${metadata[owner]}" "${metadata[file]}"
-sudo chgrp "${metadata[group]}" "${metadata[file]}"
-sudo echo "${metadata[content]}" > "${metadata[file]}"
+# sudo chmod "${metadata[permissions]}" "${metadata[file]}"
+# sudo chown "${metadata[owner]}" "${metadata[file]}"
+# sudo chgrp "${metadata[group]}" "${metadata[file]}"
+# sudo echo "${metadata[content]}" > "${metadata[file]}"
 
 # Changes permissions for all files and folders within a given directory
 # declare -A dir_metadata=(
@@ -106,6 +121,6 @@ sudo echo "${metadata[content]}" > "${metadata[file]}"
 # sudo find "${dir_metadata[directory]}" -type d -print0 | xargs -0 sudo chmod "${dir_metadata[directory_permissions]}"
 
 # --------------------------
-# Restart a Given Service
+# Check for Required Restart
 # --------------------------
 needrestart
